- refactor char form into smaller components - add model override to generate avatar - allow ltm to be disabled - better image error handling
