+++ web/pages/Character/editor.ts
@@ -21,7 +21,7 @@ import { ImageSettings } from '/common/types/image-schema'
 import { useImageCache } from '/web/shared/hooks'
 import { imageApi } from '/web/store/data/image'
 import { v4 } from 'uuid'
-import { forms } from '/web/emitter'
+import { forms, getFormValue } from '/web/emitter'
 import { ResponseSchema } from '/common/types/library'
 
 type CharKey = keyof NewCharacter
@@ -285,7 +285,8 @@ export function useCharEditor(editing?: NewCharacter & { _id?: string }) {
     const current = payload()
     const attributes = getAttributeMap(form())
     const desc = current.appearance || (attributes?.appeareance || attributes?.looks)?.join(', ')
-    const avatar = await generateAvatar(desc || '')
+    const override = getFormValue('imageOverride')
+    const avatar = await generateAvatar(desc || '', override)
     if (!avatar) return
 
     return receiveAvatar(avatar)
@@ -508,7 +509,7 @@ function getPayload(ev: any, state: EditState, original?: NewCharacter) {
   return payload
 }
 
-async function generateAvatar(description: string) {
+async function generateAvatar(description: string, override?: string) {
   const { user } = userStore.getState()
   if (!user) {
     return toastStore.error(`Image generation settings missing`)
@@ -518,7 +519,7 @@ async function generateAvatar(description: string) {
   // return image
 
   return new Promise<File>((resolve, reject) => {
-    characterStore.generateAvatar(user, description, (err, image) => {
+    characterStore.generateAvatar({ user, persona: description, override }, (err, image) => {
       if (image) return resolve(image)
       reject(err)
     })
